Update to include second pin read
This update includes a second pin read to incorporate use of a second main-loop functionality based upon a second switch. Input-1 still is responsible for initializing the Twinkles, but input-2 is now capable of stealing the entire main loop for a seperate sequence (in this case setting all LEDs to blue on WS2801 strips).